+++ frontstage/static/.well-known/security.txt
@@ -1,2 +1,2 @@
 Contact: https://hackerone.com/52fa7bc0-5356-4c86-9f79-eeb03e1d55cc/embedded_submissions/new 
-Policy: https://surveys.ons.gov.uk/help/disclosure
+Policy: https://surveys.ons.gov.uk/help/vulnerability-reporting
